BackEnd/requirements.txt
django
uwsgi
django-cors-headers
djangorestframework
libtmux
django-apscheduler